Make the Hero match designs
